ai_interview/ai/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
PyPDF2
pydantic>=2.0.0
pymongo